renamed general_model.py to xrsd_model.py and refactored XRSDModel class
